Restruturação dos Dtos e Serviços

+++ Assembly.Projecto.Final.Services/Services/AccountService.cs
@@ -1,6 +1,8 @@
-﻿using Assembly.Projecto.Final.Domain.Core.Repositories;
+﻿using Assembly.Projecto.Final.Domain.Common;
+using Assembly.Projecto.Final.Domain.Core.Repositories;
 using Assembly.Projecto.Final.Domain.Models;
 using Assembly.Projecto.Final.Services.Dtos;
+using Assembly.Projecto.Final.Services.Dtos.IServiceDtos.OtherModelsDtos;
 using Assembly.Projecto.Final.Services.Interfaces;
 using AutoMapper;
 using System;
@@ -14,39 +16,126 @@ namespace Assembly.Projecto.Final.Services.Services
     public class AccountService : IAccountService
     {
         private readonly IUnitOfWork _unitOfWork;
+        private readonly IMapper _mapper;
 
         public AccountService(IUnitOfWork unitOfWork) 
         {
             _unitOfWork = unitOfWork;
         }
-        public Account Add(Account account)
+        public AccountDto Add(CreateAccountDto createAccountDto)
         {
-            return _unitOfWork.AccountRepository.Add(account);
+            var account =_mapper.Map<Account>(createAccountDto);
+
+            Account addedAccount; 
+
+            using (_unitOfWork) 
+            {
+                _unitOfWork.BeginTransaction();
+
+                 addedAccount = _unitOfWork.AccountRepository.Add(account);
+
+                _unitOfWork.Commit();
+            }
+
+            return _mapper.Map<AccountDto>(addedAccount);
         }
 
-        public Account Delete(Account account)
+        public AccountDto Delete(AccountDto accountDto)
         {
-             return _unitOfWork.AccountRepository.Delete(account);
+             var account = _mapper.Map<Account>(accountDto);
+
+            Account deletedAccount;
+
+            using(_unitOfWork) 
+            {
+                _unitOfWork.BeginTransaction();
+
+                var foundedAccount = _unitOfWork.AccountRepository.GetById(account.Id);
+
+                if (foundedAccount == null) 
+                {
+                    throw new ArgumentNullException(nameof(foundedAccount), "Not found");
+                }
+
+                deletedAccount = _unitOfWork.AccountRepository.Delete(foundedAccount);
+
+                _unitOfWork.Commit();
+            }
+
+             return _mapper.Map<AccountDto>(deletedAccount);
         }
 
-        public Account? Delete(int id)
+        public AccountDto Delete(int id)
         {
-            return _unitOfWork.AccountRepository.Delete(id);
+
+            Account deletedAccount;
+
+            using (_unitOfWork) 
+            {
+                _unitOfWork.BeginTransaction();
+
+                var foundedAccount = _unitOfWork.AccountRepository.GetById(id);
+
+                if(foundedAccount == null) 
+                {
+                    throw new ArgumentNullException(nameof(foundedAccount), "Not found");
+                }
+
+                deletedAccount = _unitOfWork.AccountRepository.Delete(id);
+
+                _unitOfWork.Commit();
+            }
+                
+
+            return _mapper.Map<AccountDto>(deletedAccount);
         }
 
-        public List<Account> GetAll()
+        public List<AccountDto> GetAll()
         {
-            return _unitOfWork.AccountRepository.GetAll();
+            var list = new List<AccountDto>();
+
+            foreach(var account in _unitOfWork.AccountRepository.GetAll()) 
+            {
+                var accountDto = _mapper.Map<AccountDto>(account);
+
+                list.Add(accountDto);
+            }
+
+            return list;
         }
 
-        public Account? GetById(int id)
+        public AccountDto GetById(int id)
         {
-            return _unitOfWork.AccountRepository.GetById(id);
+            var account = _unitOfWork.AccountRepository.GetById(id);
+
+            return _mapper.Map<AccountDto>(account);
         }
 
-        public Account Update(Account account)
+        public AccountDto Update(AccountDto accountDto)
         {
-            return _unitOfWork.AccountRepository.Update(account);
+            var account = _mapper.Map<Account>(accountDto);
+
+            Account updatedAccount;
+
+            using (_unitOfWork) 
+            {
+                _unitOfWork.BeginTransaction();
+
+                var foundedAccount = _unitOfWork.AccountRepository.GetById(accountDto.Id);
+
+                if (foundedAccount is null)
+                {
+                    throw new ArgumentNullException(nameof(foundedAccount), "Not found");
+                }
+
+                foundedAccount.Update(accountDto.Password,accountDto.Email);
+
+                updatedAccount = _unitOfWork.AccountRepository.Update(foundedAccount);
+
+                _unitOfWork.Commit();
+            } 
+
+            return _mapper.Map<AccountDto>(updatedAccount);
         }
     }
 }
